Organize src into subpackages
