+++ src/README.md
@@ -2,5 +2,9 @@
 
 Reusable code modules for data loading, preprocessing, feature engineering, and modeling.
 
-- Organize by functionality (e.g., `data/`, `features/`, `models/`).
-- Add `__init__.py` files to make subfolders packages.
+The source code is organized into subpackages:
+
+- `data/` - Loading and preprocessing utilities
+- `analysis/` - Analysis helpers such as the correlation analyzer
+
+Each subfolder contains an `__init__.py` so they can be imported as packages.
